+++ packages/ts-moose-lib/src/dmv2/sdk/consumptionApi.ts
@@ -1,0 +1,89 @@
+import { IJsonSchemaCollection } from "typia";
+import { TypedBase } from "../typedBase";
+import { Column } from "../../dataModels/dataModelTypes";
+import { getMooseInternal } from "../internal";
+import { ConsumptionUtil } from "../../consumption-apis/helpers";
+
+/**
+ * Defines the signature for a handler function used by a Consumption API.
+ * @template T The expected type of the request parameters or query parameters.
+ * @template R The expected type of the response data.
+ * @param params An object containing the validated request parameters, matching the structure of T.
+ * @param utils Utility functions provided to the handler, e.g., for database access (`runSql`).
+ * @returns A Promise resolving to the response data of type R.
+ */
+type ConsumptionHandler<T, R> = (
+  params: T,
+  utils: ConsumptionUtil,
+) => Promise<R>;
+
+/**
+ * @template T The data type of the request parameters.
+ */
+export interface EgressConfig<T> {
+  /**
+   * An optional version string for this configuration.
+   */
+  version?: string;
+  metadata?: { description?: string };
+}
+
+/**
+ * Represents a Consumption API endpoint (Egress API), used for querying data from a Moose system.
+ * Exposes data, often from an OlapTable or derived through a custom handler function.
+ *
+ * @template T The data type defining the expected structure of the API's query parameters.
+ * @template R The data type defining the expected structure of the API's response body. Defaults to `any`.
+ */
+export class ConsumptionApi<T, R = any> extends TypedBase<T, EgressConfig<T>> {
+  metadata?: { description?: string };
+  /** @internal The handler function that processes requests and generates responses. */
+  _handler: ConsumptionHandler<T, R>;
+  /** @internal The JSON schema definition for the response type R. */
+  responseSchema: IJsonSchemaCollection.IV3_1;
+
+  /**
+   * Creates a new ConsumptionApi instance.
+   * @param name The name of the consumption API endpoint.
+   * @param handler The function to execute when the endpoint is called. It receives validated query parameters and utility functions.
+   * @param config Optional configuration for the consumption API.
+   */
+  constructor(name: string, handler: ConsumptionHandler<T, R>, config?: {});
+
+  /** @internal **/
+  constructor(
+    name: string,
+    handler: ConsumptionHandler<T, R>,
+    config: EgressConfig<T>,
+    schema: IJsonSchemaCollection.IV3_1,
+    columns: Column[],
+    responseSchema: IJsonSchemaCollection.IV3_1,
+  );
+
+  constructor(
+    name: string,
+    handler: ConsumptionHandler<T, R>,
+    config?: EgressConfig<T>,
+    schema?: IJsonSchemaCollection.IV3_1,
+    columns?: Column[],
+    responseSchema?: IJsonSchemaCollection.IV3_1,
+  ) {
+    super(name, config ?? {}, schema, columns);
+    this.metadata = config?.metadata;
+    this._handler = handler;
+    this.responseSchema = responseSchema ?? {
+      version: "3.1",
+      schemas: [{ type: "array", items: { type: "object" } }],
+      components: { schemas: {} },
+    };
+    getMooseInternal().egressApis.set(name, this);
+  }
+
+  /**
+   * Retrieves the handler function associated with this Consumption API.
+   * @returns The handler function.
+   */
+  getHandler = (): ConsumptionHandler<T, R> => {
+    return this._handler;
+  };
+}
